feat: changing inventory selection with nums and scroll

+++ include/player.h
@@ -93,7 +93,7 @@ typedef struct {
     float *z_buffer;
     hand_inv_t *hand_inv;
     sfText *visor;
-    size_t cursor;
+    int cursor;
 } player_t;
 
 #endif /* PLAYER_H_ */
